[refs #85843] More refactoring
