+++ docker-compose.yml
@@ -18,9 +18,6 @@ services:
   mysql:
     image: mysql:5.5
     container_name: fcs.db
-
-mysql:
-    image: mysql:5.5
     environment:
     - MYSQL_DATABASE=fcs
     - MYSQL_USER=fcs
